#69: hello_world.rs works with printing and exit code
