+++ jodin-rs-vm/Cargo.toml
@@ -22,3 +22,4 @@ lazy_static = "1.4"
 simplelog = { version = "0.11.0", features=["paris"] }
 log = "0.4.14"
 paris = "1.5.8"
+thiserror = "1.0.30"
